fix(QueryEditor): Ensure the correct fields appear for the given query selected
